+++ src/main/java/com/yichao/alexa/expertreview/SessionConstant.java
@@ -1,0 +1,12 @@
+package com.yichao.alexa.expertreview;
+
+public final class SessionConstant {
+
+    private SessionConstant() {
+    }
+
+    public static final String SESSION_SEARCHED_PRODUCT = "session_searched_product";
+    public static final String SESSION_LAST_RESPONSE = "session_last_response";
+    public static final String SESSION_SEARCH_RESULTS = "session_search_results";
+    public static final String SESSION_PROMPTED_SEARCH_RESULT = "session_prompted_search_result";
+}
