+++ src/test/java/com/yichao/alexa/http/parser/BaseIntegrationTest.java
@@ -1,20 +1,14 @@
 package com.yichao.alexa.http.parser;
 
-import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
-import com.yichao.alexa.http.client.CnetPageClient;
+import com.yichao.alexa.expertreview.ExpertReviewModule;
+import com.yichao.alexa.expertreview.IntentRequestHandlerModule;
 import org.junit.Before;
 
 public abstract class BaseIntegrationTest {
 
-    protected Injector injector = Guice.createInjector(new AbstractModule() {
-        @Override
-        protected void configure() {
-            bind(CnetPageClient.class);
-            bind(CnetSearchResultPageParser.class);
-        }
-    });
+    protected Injector injector = Guice.createInjector(new ExpertReviewModule(), new IntentRequestHandlerModule());
 
     @Before
     public void setup() {
